requirements.txt
langchain
PyPDF2
python-dotenv
streamlit
openai
altair
tiktoken
regex
redis
